Add extraction lambda param

+++ colbert/infra/config/settings.py
@@ -168,7 +168,7 @@ class TrainingSettings:
 
     return_max_scores: bool = DefaultVal(False)
 
-    extractions_lambda: int = DefaultVal(0.5)
+    extractions_lambda: float = DefaultVal(0.5)
 
 
 @dataclass
